Some refactoring, smash tons of build warnings, fix release build - Also make enemies not collide with each other, this resulted in exhaustion of the cached free contacts

+++ Source/CommonUtilities/Timer.cpp
@@ -26,7 +26,7 @@ namespace CommonUtilities
 		myTotalTime = fl_Duration.count();
 
 		myDeltaStart = std::chrono::high_resolution_clock::now();
-		std::chrono::duration<double> db_Duration = myDeltaStart - myDeltaEnd;
+		std::chrono::duration<float> db_Duration = myDeltaStart - myDeltaEnd;
 		myDeltaTime = db_Duration.count();
 		myDeltaEnd = myDeltaStart;
 	}
