+++ Source/Game/CharacterAnimator.cpp
@@ -43,7 +43,7 @@ void CharacterAnimator::SetDirection(float aDirection)
 {
 	assert(aDirection != 0.0f);
 
-	myDirection = MathHelper::Signum(aDirection);
+	myDirection = static_cast<float>(MathHelper::Signum(aDirection));
 }
 
 float CharacterAnimator::GetDirection() const
